# docs/openapi.yaml
openapi: 3.1.0
info:
  title: HelpDesk API (AS-IS)
  version: 0.1.0
  description: OpenAPI baseline reflecting currently implemented routes.
servers:
  - url: https://example.com
    description: Replace with deployment host
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  schemas:
    ErrorAsIs:
      type: object
      properties:
        error:
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
...
    TeamSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        organizationId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        memberCount:
          type: integer
          description: Number of team members
        activeTicketCount:
          type: integer
          description: Active tickets assigned to this team
      additionalProperties: true
...
    TeamCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
    TeamUpdate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
...
paths:
...
  /api/admin/teams:
    get:
      summary: List teams
      description: Returns all teams in the admin's organization. Requires ADMIN role.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Teams list
          content:
            application/json:
              schema:
                type: object
                properties:
                  teams:
                    type: array
                    items:
                      $ref: '#/components/schemas/TeamSummary'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
        '403':
          description: Forbidden (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
    post:
      summary: Create team
      description: Creates a new team in the admin's organization. Requires ADMIN role.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamCreate'
      responses:
        '201':
          description: Team created
          content:
            application/json:
              schema:
                type: object
                properties:
                  team:
                    $ref: '#/components/schemas/TeamSummary'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
        '403':
          description: Forbidden (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
  /api/admin/teams/{id}:
    get:
      summary: Get team
      description: Returns a specific team. Requires ADMIN role and team must be in admin's organization.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Team details
          content:
            application/json:
              schema:
                type: object
                properties:
                  team:
                    $ref: '#/components/schemas/TeamSummary'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
        '403':
          description: Forbidden (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
        '404':
          description: Team not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
    patch:
      summary: Update team
      description: Updates a team in the admin's organization. Requires ADMIN role.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamUpdate'
      responses:
        '200':
          description: Team updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  team:
                    $ref: '#/components/schemas/TeamSummary'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
        '403':
          description: Forbidden (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
        '404':
          description: Team not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
    delete:
      summary: Delete team
      description: Deletes a team from the admin's organization. Requires ADMIN role.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Team deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
        '403':
          description: Forbidden (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
        '404':
          description: Team not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
...